requirements.txt
Flask==2.3.2
mongoengine==0.27.0
flask-login==0.6.2
Werkzeug==2.3.6
requests==2.31.0
beautifulsoup4==4.12.2
google-search-results==2.4.2
python-dotenv==1.0.0
Pillow==9.5.0
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.1
statsmodels==0.13.5
deap==1.3.3
plotly==5.15.0
inflect==5.3.0
torch==2.0.1

